requirements.txt
python-dotenv~=0.20.0
python-telegram-bot~=13.12
requests~=2.28.0
configargparse~=1.5.3
Pillow~=9.1.1